Changing from models to dal to be more consistent with new impl of sql persistence
