+++ src/dal/mod.rs
@@ -1,0 +1,5 @@
+mod sql;
+mod wallet;
+use sql::{SQLConnectionType, SQL_CONNECTION};
+
+use wallet::WalletRecord;
