+++ src/lib.rs
@@ -1,4 +1,5 @@
-mod models;
+mod dal;
+mod protocols;
 mod services;
 mod tasks;
