+++ src/tasks/node.rs
@@ -1,3 +1,4 @@
+use crate::services::{insecure_testnet_keygen, Storage};
 use smol::lock::RwLock;
 use smol::net::SocketAddr;
 use smol::Timer;
